+++ src/entry/dependenciesNumber/dependenciesLOG2E.generated.ts
@@ -1,9 +1,0 @@
-/**
- * THIS FILE IS AUTO-GENERATED
- * DON'T MAKE CHANGES HERE
- */
-import { createLOG2E } from '../../factoriesNumber.js'
-
-export const LOG2EDependencies: Record<string, unknown> = {
-  createLOG2E
-}