+++ src/entry/dependenciesNumber/dependenciesParseNumberWithConfig.generated.ts
@@ -1,9 +1,0 @@
-/**
- * THIS FILE IS AUTO-GENERATED
- * DON'T MAKE CHANGES HERE
- */
-import { createParseNumberWithConfig } from '../../factoriesNumber.js'
-
-export const parseNumberWithConfigDependencies: Record<string, unknown> = {
-  createParseNumberWithConfig
-}