+++ src/entry/dependenciesNumber/dependenciesSubsetTransform.generated.ts
@@ -1,9 +1,0 @@
-/**
- * THIS FILE IS AUTO-GENERATED
- * DON'T MAKE CHANGES HERE
- */
-import { createSubsetTransform } from '../../factoriesNumber.js'
-
-export const subsetTransformDependencies: Record<string, unknown> = {
-  createSubsetTransform
-}