+++ src/plain/number/constants.ts
@@ -1,4 +1,0 @@
-export const pi: number = Math.PI
-export const tau: number = 2 * Math.PI
-export const e: number = Math.E
-export const phi: number = 1.6180339887498948 // eslint-disable-line no-loss-of-precision