+++ test/browser-test-config/browser-tests.test.ts
@@ -1,4 +1,0 @@
-// @ts-nocheck
-const testsContext = require.context('../unit-tests/', true, /.test\.js$/)
-
-testsContext.keys().forEach(testsContext)