+++ test/node-tests/treeShaking/treeShakingApp.ts
@@ -1,4 +1,0 @@
-// @ts-nocheck
-import { abs } from '../../../lib/esm/index.js'
-
-console.log(abs(-3))