+++ test/unit-tests/core/index.test.ts
@@ -1,2 +1,0 @@
-// @ts-nocheck
-// TODO: test core