+++ test/unit-tests/expression/security.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert'
-import math from '../../../src/defaultInstance.ts'
+import math from '../../../src/defaultInstance.js'
 
 describe('security', function () {
   it('should not allow calling Function via constructor', function () {